Moved GitHashHelper and PrintHelper to doctools/output.py module
